% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_shared_feature.R
\name{assign_shared_feature}
\alias{assign_shared_feature}
\title{assign_shared_feature}
\usage{
assign_shared_feature(so, min_edge_prop = 0.4)
}
\arguments{
\item{so}{a SiFINeT object}

\item{min_edge_prop}{minimum proportion of edges between a gene and a unique feature gene set for the new gene to be assigned to the set}
}
\value{
SiFINeT object with shared feature genes in featureset updated.
}
\description{
The function assigns non-unique candidate feature genes as shared feature genes into unique feature gene sets
}
\details{
Candidate feature genes that are not chosen as unique feature genes would be reconsidered as shared feature genes. 
A non-unique candidate feature gene would be assigned to a unique feature gene group if it is connected to more than min_edge_prop of the unique genes in the group.
}
