% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_coexp_sp}
\alias{cal_coexp_sp}
\title{cal_coexp_sp
This function calculates the coexpression patterns between genes
in sparse matrix and returns the coexpression matrix.}
\usage{
cal_coexp_sp(X)
}
\arguments{
\item{X}{Input binarized cell (row) by gene (column) sparse matrix}
}
\value{
Coexpression matrix
}
\description{
cal_coexp_sp
This function calculates the coexpression patterns between genes
in sparse matrix and returns the coexpression matrix.
}
\author{
Qi Gao
}
