% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SiFINeT_object.R
\name{create_SiFINeT_object}
\alias{create_SiFINeT_object}
\title{create_SiFINeT_object}
\usage{
create_SiFINeT_object(
  counts,
  gene.name = NULL,
  meta.data = NULL,
  data.name = NULL,
  sparse = FALSE,
  rowfeature = TRUE
)
}
\arguments{
\item{counts}{count matrix}

\item{gene.name}{name of the features}

\item{meta.data}{data.frame of meta data}

\item{data.name}{name of dataset}

\item{sparse}{whether the count matrix should be analyzed as sparse matrix}

\item{rowfeature}{whether the count matrix is feature (row) by cell (column)}
}
\value{
a SiFINeT object
}
\description{
The function classifies count data based on thresholds 
defined by quantile regression
}
