% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_feature_set.R
\name{enrich_feature_set}
\alias{enrich_feature_set}
\title{enrich_feature_set}
\usage{
enrich_feature_set(so, min_edge_prop = 0.9)
}
\arguments{
\item{so}{a SiFINeT object}

\item{min_edge_prop}{minimum proportion of edges between a gene and a unique+shared feature gene set for the new feature to be assigned to the set}
}
\value{
SiFINeT object with enriched feature genes in geneset updated.
}
\description{
The function chooses genes that are not found to be feature genes as enriched feature genes and assigns them into unique+shared feature gene sets
}
\details{
Genes that are not selected as feature genes would be added in the enriched section of the feature gene set if they are connected with 
more than min_edge_prop of the unique and shared feature genes in each of the feature gene group.
}
