% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstNull}
\alias{EstNull}
\title{EstNull
This function is a Rcpp version of Wenguang Sun and Tony T. Cai's 
EstNull.func R function, estimating null distribution from data.
Wenguang Sun and Tony T. Cai. "Oracle and adaptive compound decision rules for false discovery rate control". 
In: Journal of the American Statistical Association 102 (479 2007), pp. 901–912. doi: 10.1198/016214507000000545.}
\usage{
EstNull(x, gamma = 0.1)
}
\arguments{
\item{x}{Input vector of all coexpression values}

\item{gamma}{Parameter setting the stopping threshold}
}
\value{
List of mean and std
}
\description{
EstNull
This function is a Rcpp version of Wenguang Sun and Tony T. Cai's 
EstNull.func R function, estimating null distribution from data.
Wenguang Sun and Tony T. Cai. "Oracle and adaptive compound decision rules for false discovery rate control". 
In: Journal of the American Statistical Association 102 (479 2007), pp. 901–912. doi: 10.1198/016214507000000545.
}
\author{
Qi Gao
}
