% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_coexp.R
\name{feature_coexp}
\alias{feature_coexp}
\title{feature_coexp}
\usage{
feature_coexp(so)
}
\arguments{
\item{so}{a SiFINeT object}
}
\value{
SiFINeT object with coexp (gene coexpression matrix) updated.
}
\description{
The function calculates coexpression patterns between genes
}
\details{
The coexpression pattern of a pair of genes is a normalized co-occurrence of high (or equivalently low) expression level of the 2 genes in the classified count matrix.
The normalization is based on the estimated quantiles of the low-high separation instead of the quantiles used for quantile regressions.
Theoretically, the distribution of coexpression patterns should asymptotically follow standard Gaussian distribution if at least one of the 2 genes is not differentially expressed feature gene.
}
