\name{SiFINeT-package}
\alias{SiFINeT-package}
\alias{SiFINeT}
\docType{package}
\title{\packageTitle{SiFINeT}}
\description{\packageDescription{SiFINeT}}
\section{Package Content}{\packageIndices{SiFINeT}}
\author{\packageAuthor{SiFINeT}}
\section{Maintainer}{\packageMaintainer{SiFINeT}}
\keyword{package}
